backend/requirements.txt
accelerate==1.10.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cors==1.0.1
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
gevent==25.8.2
greenlet==3.2.4
hf-xet==1.1.10
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
psutil==7.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-file==2.1.0
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tldextract==5.3.0
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
zope.event==6.0
zope.interface==8.0
clip @ git+https://github.com/openai/CLIP.git